rename degradation parameters --> catalytic_rates and michaelis_constants, add flagella_expression test
