+++ vivarium/plots/gene_expression.py
@@ -1,0 +1,72 @@
+from __future__ import absolute_import, division, print_function
+
+import os
+
+import numpy as np
+import matplotlib.pyplot as plt
+
+
+def plot_timeseries_heatmaps(timeseries, config, out_dir='out'):
+    ''' make a timeseries heatmap for each port specified in config['plot_ports'] '''
+
+    name = config.get('name', 'timeseries')
+    plot_ports = config.get('plot_ports', {})
+    ports = config.get('ports', {})
+    time = timeseries['time']
+
+    def relative_to_max(series):
+        relative = max(max(series), 1)
+        return [
+            value / relative
+            for value in series]
+
+    # make timeseries heatmaps
+    ts_heatmap = {}
+    for port_id, order in plot_ports.items():
+        port = timeseries[ports[port_id]]
+        var_keys = list(order)
+
+        var_series = [
+            relative_to_max(port[key])
+            for key in var_keys]
+
+        var_keys.reverse()  # reverse to get proper labeling with imshow
+
+        ts_heatmap[port_id] = {
+            'keys': var_keys,
+            'timeseries': var_series}
+
+    # make figure for each port in plot_ports
+    for port_id, heatmap in ts_heatmap.items():
+        n_cols = 1
+        n_vars = len(heatmap['keys'])
+
+        fig = plt.figure(figsize=(4 * n_cols, 0.6 * n_vars))
+
+        var_keys = heatmap['keys']
+        var_series = heatmap['timeseries']
+        n_vars = len(var_keys)
+        ax = fig.add_subplot(111)
+
+        im = ax.imshow(var_series,
+            extent=[time[0], time[-1], 0, n_vars],
+            interpolation='nearest',
+            aspect='auto',
+            cmap='cividis'
+            )
+        ax.locator_params(axis='y', nbins=n_vars)
+
+        # set y ticks locations and labels
+        y_tick_locs = np.asarray([loc+0.5 for loc in range(n_vars)])
+        ax.set_yticks(y_tick_locs)
+        ax.set_yticklabels(var_keys)
+        ax.set_xlabel('time (s)')
+
+        # colorbar
+        cbar = fig.colorbar(im)
+        cbar.set_label('relative flourescence', rotation=270, labelpad=20)
+
+        # save figure
+        figname = name + '_' + port_id
+        fig_path = os.path.join(out_dir, figname)
+        plt.savefig(fig_path, bbox_inches='tight')
